업데이트

+++ GameEngineBase/GameEngineString.cpp
@@ -26,7 +26,7 @@ void GameEngineString::toupper(std::string& Text)
 
 void GameEngineString::StringToWString(const std::string& _Text, std::wstring& _Out)
 {
-	int Size = MultiByteToWideChar(CP_ACP, 0, _Text.c_str(), _Text.size(), nullptr, 0);
+	int Size = MultiByteToWideChar(CP_ACP, 0, _Text.c_str(), static_cast<int>(_Text.size()), nullptr, 0);
 
 	if (Size == 0)
 	{
@@ -36,7 +36,7 @@ void GameEngineString::StringToWString(const std::string& _Text, std::wstring& _
 
 	_Out.resize(Size);
 
-	Size = MultiByteToWideChar(CP_ACP, 0, _Text.c_str(), _Text.size(), &_Out[0], Size);
+	Size = MultiByteToWideChar(CP_ACP, 0, _Text.c_str(), static_cast<int>(_Text.size()), &_Out[0], Size);
 
 	if (Size == 0)
 	{
